move Tape/prettyPrint into own files
